Add IBlockStorage interface for block management
Introduced a new `IBlockStorage` interface in the `Hemelo.Database.Core.Interfaces` namespace to manage fixed-size blocks of data. The interface provides properties for block content size, header size, and total size, along with methods to find and allocate blocks. XML documentation was added to describe the interface and its members.